Setup Screen Util Code✅
- Created the `ClinicSystem` widget.
- Added `ScreenUtilInit` widget.
- Added the `font family` into `pubspec.yaml` file.
