+++ lib/clinic_system.dart
@@ -1,0 +1,34 @@
+import 'package:el_sharq_clinic/core/theming/app_colors.dart';
+import 'package:el_sharq_clinic/core/theming/app_text_styles.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
+
+class ClinicSystem extends StatelessWidget {
+  const ClinicSystem({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ScreenUtilInit(
+      designSize: const Size(1440, 1024),
+      minTextAdapt: true,
+      child: MaterialApp(
+        debugShowCheckedModeBanner: false,
+        title: 'Clinic System',
+        theme: ThemeData(
+          colorScheme: ColorScheme.fromSeed(seedColor: AppColors.blue),
+          useMaterial3: true,
+        ),
+        home: Scaffold(
+          appBar: AppBar(
+            title: Text('Clinic System',
+                style: AppTextStyles.font32DarkGreyMedium),
+          ),
+          body: Center(
+            child: Text('Welcome to Clinic System',
+                style: AppTextStyles.font28DarkGreyMedium),
+          ),
+        ),
+      ),
+    );
+  }
+}
